requirements.txt
pennylane>=0.39
torch>=2.2
rdkit-pypi>=2022.9.5
numpy>=1.24
